Refactor the code so forms and specific buttons are defined in separate files
